pyproject.toml
[project]
name = "{{project_name}}"
requires-python = ">=3.10"
authors = [{ name="{{author}}", email="{{author_email}}" }]
description = "A simple backend base"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12"
]
dynamic = ["version"]
dependencies = [
  "flask>=3.0.3",
  "flask_restx>=1.3.0",
  "gunicorn>=22.0.0",
  "gevent>=24.2.1",
  "dataclasses_json>=0.6.6",
  "requests==2.31.0",
  "python-dotenv==1.0.0"
]

[tool.setuptools]
packages = []

[project.urls]
repository = ""

[project.optional-dependencies]
test = [
  "pytest==7.4.2"
]

docs = [
  "sphinx>=8.0.2",
  "furo>=2024.8.6",
  "sphinx_copybutton>=0.5.2"
]

[tool.sphinx]
builder = "html"
